+++ lib/screen/slim/service_screen.dart
@@ -3,35 +3,38 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:letscheck/bloc/connection_data/connection_data.dart';
 import 'package:letscheck/widget/site_stats_widget.dart';
 import 'base_slim_screen.dart';
-import '../../global_router.dart';
 
 class ServiceScreen extends StatefulWidget {
-  static final route = buildRoute(
-      key: routeService,
-      uri: '/conn/{alias}/host/{hostname}/services/{service}',
-      lastArgOptional: false,
-      route: (context) => MaterialPageRoute(
-            settings: context,
-            builder: (context) => ServiceScreen(),
-          ));
+  final String alias;
+  final String hostname;
+  final String service;
+
+  ServiceScreen({
+    required this.alias,
+    required this.hostname,
+    required this.service,
+  });
 
   @override
-  ServiceScreenState createState() => ServiceScreenState();
+  ServiceScreenState createState() => ServiceScreenState(
+        alias: alias,
+        hostname: hostname,
+        service: service,
+      );
 }
 
 class ServiceScreenState extends State<ServiceScreen> with BaseSlimScreenState {
+  final String alias;
+  final String hostname;
+  final String service;
+
+  ServiceScreenState(
+      {required this.alias, required this.hostname, required this.service});
   @override
   BaseSlimScreenSettings setup(BuildContext context) {
-    final groups = ServiceScreen.route.extractNamedArgs(context);
     var title = 'Service';
-    if (!groups.containsKey('alias') ||
-        !groups.containsKey('hostname') ||
-        !groups.containsKey('service')) {
-      Navigator.of(context)
-          .pushReplacementNamed(GlobalRouter().buildUri(routeNotFound));
-    }
 
-    var serviceName = groups['service']!;
+    var serviceName = service;
     if (serviceName.length > 25) {
       serviceName = serviceName.substring(0, 25);
     }
@@ -42,11 +45,10 @@ class ServiceScreenState extends State<ServiceScreen> with BaseSlimScreenState {
 
   @override
   Widget content(BuildContext context) {
-    final groups = ServiceScreen.route.extractNamedArgs(context);
     final cBloc = BlocProvider.of<ConnectionDataBloc>(context);
     return Column(
       children: [
-        SiteStatsWidget(alias: groups['alias']!, state: cBloc.state),
+        SiteStatsWidget(alias: alias, state: cBloc.state),
         Expanded(
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.center,