+++ lib/screen/slim/settings_connection_screen.dart
@@ -1,38 +1,37 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+
 import '../../form/connection_form/connection_form.dart';
 import '../../bloc/settings/settings.dart';
 import 'base_slim_screen.dart';
-import '../../global_router.dart';
 
 class SettingsConnectionScreen extends StatefulWidget {
-  static final route = buildRoute(
-      key: routeSettingsConnection,
-      uri: '/settings/connection/{name}',
-      lastArgOptional: true,
-      route: (context) => MaterialPageRoute(
-            settings: context,
-            builder: (context) => SettingsConnectionScreen(),
-          ));
+  final String alias;
+
+  SettingsConnectionScreen({required this.alias});
 
   @override
-  SettingsConnectionScreenState createState() =>
-      SettingsConnectionScreenState();
+  SettingsConnectionScreenState createState() => SettingsConnectionScreenState(
+        alias: alias,
+      );
 }
 
 class SettingsConnectionScreenState extends State<SettingsConnectionScreen>
     with BaseSlimScreenState {
+  final String alias;
+
+  SettingsConnectionScreenState({required this.alias});
+
   @override
   BaseSlimScreenSettings setup(BuildContext context) {
-    final groups = SettingsConnectionScreen.route.extractNamedArgs(context);
     var title = 'Add Connection';
-    if (groups.containsKey('name')) {
-      title = "Connection: ${groups["name"]}";
+    if (alias != '+') {
+      title = "Connection: $alias";
     }
 
     return BaseSlimScreenSettings(
       title,
-      showLeading: groups.containsKey('name'),
+      showLeading: alias.isNotEmpty,
       showRefresh: false,
       showMenu: false,
       showSettings: false,
@@ -42,13 +41,9 @@ class SettingsConnectionScreenState extends State<SettingsConnectionScreen>
 
   @override
   Widget content(BuildContext context) {
-    final groups = SettingsConnectionScreen.route.extractNamedArgs(context);
-
     final sBloc = BlocProvider.of<SettingsBloc>(context);
 
-    if (groups.containsKey('name')) {
-      var alias = groups['name']!;
-
+    if (alias != '+') {
       return BlocProvider(
         create: (context) => ConnectionFormBloc(
           settingsBloc: sBloc,