+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - upgrade flutter 3.27.1 -> 3.27.3
 - redirect home screen to add connection when theres no connection yet
+- replace howmegrown router with go\_router
 
 ## 0.1.3+1269 - 2025-02-05
 ### Added